requirements.txt
azure-ai-agents
azure-identity
azure-ai-projects